requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
atomicwrites==1.4.1
attrs==25.3.0
beautifulsoup4==4.9.3
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.8.3
charset-normalizer==2.0.12
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Deprecated==1.2.18
deprecation==2.1.0
docopt==0.6.2
Flask==3.1.1
flask-supabase==0.2.1
gotrue==2.12.3
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
lxml==6.0.0
MarkupSafe==3.0.2
numpy==2.3.2
packaging==25.0
pandas==2.3.1
platformdirs==4.3.8
pluggy==1.6.0
postgrest==1.1.1
prompt_toolkit==3.0.51
py==1.11.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pytest==7.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
realtime==2.6.0
redis==4.0.2
requests==2.26.0
requests-file==2.1.0
requests-toolbelt==1.0.0
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
text-unidecode==1.3
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
zeep==4.3.1
